`active` and `request` context variables also available during building.
Required cross-updating the `hecto`library as well.
