+++ blueprint/_layout.html
@@ -8,6 +8,7 @@
   <link rel="shortcut icon" type="image/x-icon" href="/static/favicon.ico">
   <link rel="stylesheet" href="/static/normalize.css">
   <link rel="stylesheet" href="/static/styles.css?v=1.0">
+  <link type="text/plain" rel="author" href="/humans.txt">
   {% block head %}{% endblock %}
 </head>
 <body>
