+++ clay/utils/__init__.py
@@ -1,3 +1,4 @@
+from .active import *  # noqa
 from .jinja_render import *  # noqa
 from .load_config import *  # noqa
 from .path_matcher import *  # noqa
