+++ setup.cfg
@@ -21,7 +21,7 @@ packages = find:
 python_requires = >=3.6,<4.0
 install_requires =
     gevent ~= 1.4
-    hecto ~= 1.191024
+    hecto ~= 1.191026
     jinja2 ~= 2.10
     multidict ~= 4.5.2
     multipart ~= 0.2
